refactor(sdk): rename src/mcp to src/daemon-mcp
Rename the MCP utilities directory to better reflect its role as
daemon-specific MCP tooling. Update all import paths in index.ts,
Query.ts, and the bin entry in package.json.
