+++ packages/sdk-typescript/package.json
@@ -20,7 +20,7 @@
     "./package.json": "./package.json"
   },
   "bin": {
-    "qwen-serve-mcp": "./dist/mcp/serve-bridge/bin.js"
+    "qwen-serve-mcp": "./dist/daemon-mcp/serve-bridge/bin.js"
   },
   "files": [
     "dist",
