+++ packages/sdk-typescript/src/daemon-mcp/formatters.ts
