+++ packages/sdk-typescript/src/daemon-mcp/serve-bridge/README.md
