+++ packages/sdk-typescript/src/daemon-mcp/serve-bridge/bin.ts
