+++ packages/sdk-typescript/src/daemon-mcp/serve-bridge/tools/workspaceWrite.ts
