+++ packages/sdk-typescript/src/daemon-mcp/tool.ts
