+++ packages/sdk-typescript/src/index.ts
@@ -188,18 +188,18 @@ export {
 } from './daemon/index.js';
 
 // SDK MCP Server exports
-export { tool } from './mcp/tool.js';
-export { createSdkMcpServer } from './mcp/createSdkMcpServer.js';
-export { createServeBridgeMcpServer } from './mcp/serve-bridge/index.js';
+export { tool } from './daemon-mcp/tool.js';
+export { createSdkMcpServer } from './daemon-mcp/createSdkMcpServer.js';
+export { createServeBridgeMcpServer } from './daemon-mcp/serve-bridge/index.js';
 
-export type { SdkMcpToolDefinition } from './mcp/tool.js';
+export type { SdkMcpToolDefinition } from './daemon-mcp/tool.js';
 
 export type {
   CreateSdkMcpServerOptions,
   McpSdkServerConfigWithInstance,
-} from './mcp/createSdkMcpServer.js';
+} from './daemon-mcp/createSdkMcpServer.js';
 
-export type { ServeBridgeMcpServerOptions } from './mcp/serve-bridge/index.js';
+export type { ServeBridgeMcpServerOptions } from './daemon-mcp/serve-bridge/index.js';
 
 export type { QueryOptions } from './query/createQuery.js';
 export type { LogLevel, LoggerConfig, ScopedLogger } from './utils/logger.js';
